+++ site/search_script.js
@@ -343,7 +343,7 @@ const search_targets = [
 	new SearchTarget("Ephemeral Value Category", "documentation/panther/value_categories.html#ephemeral", "", ["ephemeral","value","category"], [0,4]),
 	new SearchTarget("Concrete Value Category", "documentation/panther/value_categories.html#concrete", "", ["concrete","value","category"], [0,4]),
 	new SearchTarget("Concrete-Mutable Value Category", "documentation/panther/value_categories.html#concrete_mutable", "", ["concrete-mutable","value","category"], [0,4]),
-	new SearchTarget("Concrete-Forwardable Value Category", "documentation/panther/value_categories.html#concrete_forwardable", "", ["concrete-forwardable","value","category"], [0,4]),
+	new SearchTarget("Forwardable Value Category", "documentation/panther/value_categories.html#forwardable", "", ["forwardable","value","category"], [0,4]),
 	new SearchTarget("Concrete-Const Value Category", "documentation/panther/value_categories.html#concrete_const", "", ["concrete-const","value","category"], [0,4]),
 	new SearchTarget("Value Stages", "/site/documentation/panther/value_stages.html", "Documentation for value stages in the Panther programming language", ["value","stages"], [0,4]),
 	new SearchTarget("Constexpr Value Stage", "documentation/panther/value_stages.html#constexpr", "", ["constexpr","value","stage"], [0,4]),